+++ packages/expo-glass-effect/CHANGELOG.md
@@ -9,3 +9,6 @@
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+- Move `isLiquidGlassAvailable` function to `ExpoGlassEffect` ([#39349](https://github.com/expo/expo/pull/39349) by [@Ubax](https://github.com/Ubax))
+